default

+++ part_dataset.py
@@ -92,7 +92,7 @@ class PartDataset():
         self.classes = dict(zip(self.cat, range(len(self.cat))))  
         self.num_seg_classes = 0
         if not self.classification:
-            for i in range(len(self.datapath)/50):
+            for i in range(int(len(self.datapath)/50)):
                 l = len(np.unique(np.loadtxt(self.datapath[i][-1]).astype(np.uint8)))
                 if l > self.num_seg_classes:
                     self.num_seg_classes = l
